port reconstruct to .ts & seprate style
