+++ src/style.ts
@@ -1,0 +1,47 @@
+export default `
+    <style>
+    @import url(http://fonts.googleapis.com/earlyaccess/nanummyeongjo.css);
+    body {
+        margin-top: 10px;
+        margin-bottom: 500px;
+        text-align: center;
+    }
+    #info {
+        margin-bottom: 40px;
+        color: #666;
+    }
+    #meta {
+        display: inline-block;
+        width: 640px;
+    }
+    #timestamp {
+        color: #888;
+        font-size: 10pt;
+        text-align: left;
+    }
+    #timestamp p {
+        margin: 0;
+    }
+    #reporters {
+        list-style-type: none;
+        text-align: right;
+    }
+    #reporters .mail {
+        margin-left: 8px;
+    }
+    #content {
+        display: inline-block;
+        width: 640px;
+        font-family: 'Nanum Myeongjo', serif;
+        font-size: 11pt;
+        text-align: justify;
+        line-height: 1.6;
+    }
+    #content img {
+        display: block;
+        margin: 15px auto;
+        max-width: 100%;
+        height: auto;
+    }
+    </style>
+`;
